# docs/evidence/0.6/openapi.normalized.json
{
  "components": {
    "schemas": {
      "AcceptReceiptResponse": {
        "properties": {
          "acceptance_id": {
            "title": "Acceptance Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "events_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Url"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "default": "run",
            "title": "Resource Type",
            "type": "string"
          },
          "schema": {
            "default": "etlantic.control_plane.accept_receipt/1",
            "title": "Schema",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "default": "accepted",
            "description": "Durable accept only \u2014 not pipeline execution status",
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Url"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "acceptance_id",
          "submission_id",
          "tenant_id",
          "workspace_id",
          "idempotency_key",
          "created_at"
        ],
        "title": "AcceptReceiptResponse",
        "type": "object"
      },
...
      "ArtifactMeta": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "content_available": {
            "default": false,
            "title": "Content Available",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          }
        },
        "required": [
          "artifact_id",
          "kind"
        ],
        "title": "ArtifactMeta",
        "type": "object"
      },
      "ArtifactsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactMeta"
            },
            "title": "Items",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "ArtifactsResponse",
        "type": "object"
      },
...
      "DefinitionGetResponse": {
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          }
        },
        "required": [
          "definition_id",
          "document"
        ],
        "title": "DefinitionGetResponse",
        "type": "object"
      },
...
      "DefinitionWriteBody": {
        "additionalProperties": false,
        "properties": {
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          }
        },
        "required": [
          "document"
        ],
        "title": "DefinitionWriteBody",
        "type": "object"
      },
      "DefinitionWriteResponse": {
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          }
        },
        "required": [
          "definition_id",
          "fingerprint",
          "document"
        ],
        "title": "DefinitionWriteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanRequestBody": {
        "additionalProperties": false,
        "properties": {
          "request": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request"
          },
          "revision_selector": {
            "default": "current",
            "title": "Revision Selector",
            "type": "string"
          }
        },
        "title": "PlanRequestBody",
        "type": "object"
      },
      "PlanResponse": {
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          }
        },
        "required": [
          "ok",
          "definition_id"
        ],
        "title": "PlanResponse",
        "type": "object"
      },
...
      "RunSubmitBody": {
        "description": "Optional body for run submit; Idempotency-Key header is preferred.",
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "title": "RunSubmitBody",
        "type": "object"
      },
...
      "ValidateResponse": {
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          }
        },
        "required": [
          "ok",
          "definition_id"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
...
  "paths": {
...
    "/v1/definitions/{definition_id}": {
      "get": {
        "operationId": "cp_get_definition",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Definition",
        "tags": [
          "definitions"
        ]
      },
      "put": {
        "operationId": "cp_register_definition",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinitionWriteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionWriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Definition",
        "tags": [
          "definitions"
        ]
      }
    },
...
    "/v1/definitions/{definition_id}/plan": {
      "post": {
        "operationId": "cp_plan_definition",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PlanRequestBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Definition",
        "tags": [
          "definitions"
        ]
      }
    },
...
    "/v1/definitions/{definition_id}/runs": {
      "post": {
        "operationId": "cp_submit_run",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunSubmitBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptReceiptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/definitions/{definition_id}/validate": {
      "post": {
        "operationId": "cp_validate_definition",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision_selector",
            "required": false,
            "schema": {
              "default": "current",
              "title": "Revision Selector",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Definition",
        "tags": [
          "definitions"
        ]
      }
    },
...
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "cp_get_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/artifacts": {
      "get": {
        "operationId": "cp_list_run_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Artifacts",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cp_cancel_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "description": "Stream ordered run events as ``text/event-stream``.\n\nAuthz precedes existence lookup. Cross-tenant runs map to opaque 404.\nUnknown/expired cursors return **410 Gone** (prefer reconnect without\ncursor to replay from the beginning) \u2014 see package README.",
        "operationId": "cp_stream_run_events",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Opaque resume cursor (etlantic.control_plane.sse_cursor/1)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque resume cursor (etlantic.control_plane.sse_cursor/1)",
              "title": "Cursor"
            }
          },
          {
            "description": "When true, keep polling for new events after history (capped: max 100 polls / 60s)",
            "in": "query",
            "name": "follow",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, keep polling for new events after history (capped: max 100 polls / 60s)",
              "title": "Follow",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Server-Sent Events stream of run lifecycle events"
          },
          "410": {
            "description": "Resume cursor expired or unknown; reconnect without cursor to replay from the beginning"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/lineage": {
      "get": {
        "operationId": "cp_get_run_lineage",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageStubResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Lineage",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/report": {
      "get": {
        "operationId": "cp_get_run_report",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStubResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Report",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/schema/observations/{observation_id}/ack": {
      "post": {
        "operationId": "cp_ack_schema_observation",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaObservationAckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack Schema Observation",
        "tags": [
          "schema"
        ]
      }
    },
...
  }
}